requirements.txt

coverage
Pillow==2.7.0
argparse==1.2.1
html5lib==0.999 
six==1.9.0
wsgiref==0.1.2

Django==1.7.7

django-registration-redux==1.1
django-ajax-selects==1.3.6
django-bootstrap3==5.2.0
django-classy-tags==0.6.1
django-sekizai==0.8.1
django-geoposition==0.2.2
django-recaptcha==1.0
django-social-auth==0.7.28

django-cms==3.1.3
django-treebeard==3.0
django-menus==1.1.2
djangocms-text-ckeditor==2.7.0

